+++ src/db/sql/functions/helper/get_h3_3_neighbors.sql
@@ -1,15 +1,15 @@
 DROP FUNCTION basic.get_h3_neighbors;
-CREATE OR REPLACE FUNCTION basic.get_h3_neighbors(h3_3_input smallint)
- RETURNS SMALLINT[]
+CREATE OR REPLACE FUNCTION basic.get_h3_neighbors(h3_3_input integer)
+ RETURNS Integer[]
  LANGUAGE plpgsql
 AS $function$
 DECLARE 
-	h3_3_arr SMALLINT[]; 
+	h3_3_arr Integer[]; 
 BEGIN 
 	SELECT ARRAY_AGG(z.h3_3)
 	INTO h3_3_arr
 	FROM (
-		SELECT to_short_h3_3(h3_grid_ring_unsafe(to_full_from_short_h3_3(h3_3_input::smallint)::h3index)::bigint) AS h3_3
+		SELECT to_short_h3_3(h3_grid_ring_unsafe(to_full_from_short_h3_3(h3_3_input::integer)::h3index)::bigint) AS h3_3
 		UNION ALL
 		SELECT h3_3_input
 	) z;
